Actually set maxBatchSize on EventProcessor
The `max_batch_size` option was being ignored when creating the
event processor, leading to a default batch size of 10, instead of
the intended 125, and not allowing users to change their batch size.

This caused the potential for much more frequent checkpointing of
completed event batches, potentially causing performance issues.