Fix `getHostContext` method accessibility
This commit works around a recent change to jruby where calls to `java_send` to
package private accessible Java methods will fail with
`(NoMethodError) undefined method `getHostContext'` errors.

This commit works around this by using Java reflection to make the method accessible.

I have verified that this is the only package private method that we call on that class.

Fixes #91